Build: Externalize and refactor code coverage threshold
This commit refactors the code coverage configuration to use an integer-based percentage value and allows it to be overridden via an environment variable, improving flexibility across local and CI environments.

Key changes:
- **Build Logic**:
    - Changed `coverageThreshold` type from `Double` to `Int` in `CodeCoverageExtension.kt`.
    - Updated `CodeCoveragePlugin.kt` to convert the integer percentage (e.g., 80) to a decimal value (e.g., 0.80) for JaCoCo violation rules.
    - Updated `build.gradle.kts` to fetch the threshold from the `CODE_COVERAGE_THRESHOLD` environment variable, defaulting to `80`.

- **CI/CD (GitHub Actions)**:
    - Added `CODE_COVERAGE_THRESHOLD` environment variable to `feature-pull-request.yml` and `build-main.yml` workflows using GitHub repository variables.
    - Updated `code-coverage-action` to use the `${{ env.CODE_COVERAGE_THRESHOLD }}` variable for both overall and changed files coverage minimums.

+++ build-logic/code-coverage/src/main/kotlin/md/vnastasi/plugin/codecoverage/CodeCoverageExtension.kt
@@ -15,7 +15,7 @@ abstract class CodeCoverageExtension @Inject constructor(objectFactory: ObjectFa
 
     val reportDirectory: DirectoryProperty = objectFactory.directoryProperty()
 
-    val coverageThreshold: Property<Double> = objectFactory.property()
+    val coverageThreshold: Property<Int> = objectFactory.property()
 
     val excludedModules: ListProperty<Project> = objectFactory.listProperty()
